+++ node/Cargo.toml
@@ -24,27 +24,16 @@ pprof = { version = "0.6", features = ["flamegraph"] }
 bincode = "1.3.3"
 
 [[bench]]
-name = "commit_coeff_form"
+name = "commit"
+harness = false
+
+
+[[bench]]
+name = "poly_verify"
 harness = false
 
 [[bench]]
-name = "commit_eval_form"
-harness = false
-
-[[bench]]
-name = "poly_verify_coeff_form"
-harness = false
-
-[[bench]]
-name = "create_witness_coeff_form"
-harness = false
-
-[[bench]]
-name = "create_witness_eval_form"
-harness = false
-
-[[bench]]
-name = "verify_eval_coeff_form"
+name = "create_witness"
 harness = false
 
 [[bench]]
