+++ node/benches/poly_arithmetic.rs
@@ -7,33 +7,30 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 fn random_polynomial(rng: &mut SmallRng, n: usize) -> Polynomial {
     let mut coeffs = vec![Scalar::zero(); n];
-    for i in 0..n {
-        coeffs[i] = rng.gen::<u64>().into();
+    for coeff in coeffs.iter_mut().take(n) {
+        *coeff = rng.gen::<u64>().into();
     }
     Polynomial::new(coeffs)
 }
 
-fn bench_poly_arithmetic<const NUM_COEFFS: usize>(c: &mut Criterion) {
+fn poly_arithmetic<const NUM_COEFFS: usize>(c: &mut Criterion) {
     let mut rng = SmallRng::from_seed([NUM_COEFFS as u8; 32]);
     let f = random_polynomial(&mut rng, NUM_COEFFS);
     let g = random_polynomial(&mut rng, NUM_COEFFS);
 
-    c.bench_function(
-        format!("bench_add, degree {}", NUM_COEFFS - 1).as_str(),
-        |b| {
-            b.iter(|| black_box(f.clone()) + black_box(g.clone()));
-        },
-    );
+    c.bench_function(format!("add, degree {}", NUM_COEFFS - 1).as_str(), |b| {
+        b.iter(|| black_box(f.clone()) + black_box(g.clone()));
+    });
 
     c.bench_function(
-        format!("bench_mul_naive, degree {}", NUM_COEFFS - 1).as_str(),
+        format!("mul_naive, degree {}", NUM_COEFFS - 1).as_str(),
         |b| {
             b.iter(|| black_box(f.clone()) * black_box(g.clone()));
         },
     );
 
     c.bench_function(
-        format!("bench_mul_fft, degree {}", NUM_COEFFS - 1).as_str(),
+        format!("mul_fft, degree {}", NUM_COEFFS - 1).as_str(),
         |b| {
             b.iter(|| black_box(f.clone()).fft_mul(black_box(&g)));
         },
@@ -42,7 +39,7 @@ fn bench_poly_arithmetic<const NUM_COEFFS: usize>(c: &mut Criterion) {
     let g = random_polynomial(&mut rng, 2);
 
     c.bench_function(
-        format!("bench_long_division, degree {}", NUM_COEFFS - 1).as_str(),
+        format!("long_division, degree {}", NUM_COEFFS - 1).as_str(),
         |b| b.iter(|| black_box(f.clone()).long_division(&black_box(g.clone()))),
     );
 
@@ -54,14 +51,14 @@ fn bench_poly_arithmetic<const NUM_COEFFS: usize>(c: &mut Criterion) {
     }
 
     c.bench_function(
-        format!("bench_interpolation, degree {}", NUM_COEFFS - 1).as_str(),
+        format!("interpolation, degree {}", NUM_COEFFS - 1).as_str(),
         |b| b.iter(|| Polynomial::lagrange_interpolation(xs.as_slice(), ys.as_slice())),
     );
 }
 
 criterion_group!(
-    name = poly_arithmetic;
+    name = bench_poly_arithmetic;
     config = Criterion::default();
-    targets = bench_poly_arithmetic<16>, bench_poly_arithmetic<64>, bench_poly_arithmetic<128>, bench_poly_arithmetic<256>, bench_poly_arithmetic<512>, bench_poly_arithmetic<1024>, bench_poly_arithmetic<2048>
+    targets = poly_arithmetic<16>, poly_arithmetic<64>, poly_arithmetic<128>, poly_arithmetic<256>, poly_arithmetic<512>, poly_arithmetic<1024>, poly_arithmetic<2048>, poly_arithmetic<5098>
 );
-criterion_main!(poly_arithmetic);
+criterion_main!(bench_poly_arithmetic);
