+++ node/benches/poly_verify.rs
@@ -1,8 +1,7 @@
 use blstrs::Scalar;
 use melon::kzg::polynomial::Polynomial;
 use melon::kzg::{
-    coeff_form::{KZGProver, KZGVerifier},
-    setup, KZGParams,
+    setup, KZGParams, {KZGProver, KZGVerifier},
 };
 use pairing::group::ff::Field;
 use rand::{rngs::SmallRng, Rng, SeedableRng};
@@ -14,12 +13,12 @@ pub fn csprng_setup<const MAX_COEFFS: usize>() -> KZGParams {
     setup(s, MAX_COEFFS)
 }
 
-fn bench_poly_verify<const NUM_COEFFS: usize>(c: &mut Criterion) {
+fn poly_verify<const NUM_COEFFS: usize>(c: &mut Criterion) {
     let params = csprng_setup::<NUM_COEFFS>();
     let mut rng = SmallRng::from_seed([42; 32]);
     let mut coeffs = vec![Scalar::zero(); NUM_COEFFS];
-    for i in 0..NUM_COEFFS {
-        coeffs[i] = rng.gen::<u64>().into();
+    for coeff in coeffs.iter_mut().take(NUM_COEFFS) {
+        *coeff = rng.gen::<u64>().into();
     }
     let polynomial = Polynomial::new_from_coeffs(coeffs, NUM_COEFFS - 1);
     let prover = KZGProver::new(&params);
@@ -27,7 +26,7 @@ fn bench_poly_verify<const NUM_COEFFS: usize>(c: &mut Criterion) {
     let commitment = prover.commit(&polynomial);
 
     c.bench_function(
-        format!("bench_verify_poly_coeff_form, degree {}", NUM_COEFFS - 1).as_str(),
+        format!("verify_poly, degree {}", NUM_COEFFS - 1).as_str(),
         |b| {
             b.iter(|| {
                 verifier.verify_poly(black_box(&commitment), black_box(&polynomial));
@@ -37,8 +36,8 @@ fn bench_poly_verify<const NUM_COEFFS: usize>(c: &mut Criterion) {
 }
 
 criterion_group!(
-    name = poly_verify;
+    name = bench_poly_verify;
     config = Criterion::default();
-    targets = bench_poly_verify<16>, bench_poly_verify<64>, bench_poly_verify<128>, bench_poly_verify<256>, bench_poly_verify<512>, bench_poly_verify<1024>, bench_poly_verify<2048>
+    targets = poly_verify<16>, poly_verify<64>, poly_verify<128>, poly_verify<256>, poly_verify<512>, poly_verify<1024>, poly_verify<2048>, poly_verify<5096>
 );
-criterion_main!(poly_verify);
+criterion_main!(bench_poly_verify);
