+++ node/src/kzg/mod.rs
@@ -1,12 +1,19 @@
-use blstrs::{G1Affine, G1Projective, G2Projective, Scalar};
+use blstrs::{pairing, G1Affine, G1Projective, G2Projective, Scalar};
 use pairing::group::Group;
 use thiserror::Error;
 
-pub mod coeff_form;
-pub mod eval_form;
 pub mod ft;
 pub mod polynomial;
-pub mod utils;
+
+use polynomial::Polynomial;
+
+use pairing::group::{ff::Field, prime::PrimeCurveAffine, Curve};
+use std::fmt::Debug;
+
+#[derive(Debug, Clone)]
+pub struct KZGProver<'params> {
+    parameters: &'params KZGParams,
+}
 
 /// parameters from tested setup
 #[derive(Clone, Debug)]
@@ -53,11 +60,239 @@ pub fn setup(s: Scalar, num_coeffs: usize) -> KZGParams {
     KZGParams { gs, hs }
 }
 
-#[cfg(any(csprng_setup, test))]
-use rand::random;
+#[derive(Debug, Clone)]
+pub struct KZGVerifier<'params> {
+    parameters: &'params KZGParams,
+}
 
-#[cfg(any(csprng_setup, test))]
-pub fn csprng_setup(num_coeffs: usize) -> KZGParams {
-    let s: Scalar = random::<u64>().into();
-    setup(s, num_coeffs)
+impl<'params> KZGProver<'params> {
+    /// initializes `polynomial` to zero polynomial
+    pub fn new(parameters: &'params KZGParams) -> Self {
+        Self { parameters }
+    }
+
+    pub fn parameters(&self) -> &'params KZGParams {
+        self.parameters
+    }
+
+    pub fn commit(&self, polynomial: &Polynomial) -> KZGCommitment {
+        let gs = &self.parameters.gs[..polynomial.num_coeffs()];
+        let commitment = G1Projective::multi_exp(gs, polynomial.slice_coeffs());
+
+        commitment.to_affine()
+    }
+
+    pub fn create_witness(
+        &self,
+        polynomial: &Polynomial,
+        (x, y): (Scalar, Scalar),
+    ) -> Result<KZGWitness, KZGError> {
+        let mut dividend = polynomial.clone();
+        dividend.coeffs[0] -= y;
+
+        let divisor = Polynomial::new_from_coeffs(vec![-x, Scalar::one()], 1);
+        match dividend.long_division(&divisor) {
+            // by polynomial remainder theorem, if (x - point.x) does not divide self.polynomial, then
+            // self.polynomial(point.y) != point.1
+            (_, Some(_)) => Err(KZGError::PointNotOnPolynomial),
+            (psi, None) if psi.num_coeffs() == 1 => {
+                Ok((self.parameters.gs[0] * psi.coeffs[0]).to_affine())
+            }
+            (psi, None) => {
+                let gs = &self.parameters.gs[..psi.num_coeffs()];
+                Ok(G1Projective::multi_exp(gs, psi.slice_coeffs()).to_affine())
+            }
+        }
+    }
+}
+
+impl<'params> KZGVerifier<'params> {
+    pub fn new(parameters: &'params KZGParams) -> Self {
+        KZGVerifier { parameters }
+    }
+
+    pub fn verify_poly(&self, commitment: &KZGCommitment, polynomial: &Polynomial) -> bool {
+        let gs = &self.parameters.gs[..polynomial.num_coeffs()];
+        let check = G1Projective::multi_exp(gs, polynomial.slice_coeffs());
+
+        check.to_affine() == *commitment
+    }
+
+    pub fn verify_eval(
+        &self,
+        (x, y): (Scalar, Scalar),
+        commitment: &KZGCommitment,
+        witness: &KZGWitness,
+    ) -> bool {
+        let lhs = pairing(
+            witness,
+            &(self.parameters.hs[1] - self.parameters.hs[0] * x).to_affine(),
+        );
+        let rhs = pairing(
+            &(commitment.to_curve() - self.parameters.gs[0] * y).to_affine(),
+            &self.parameters.hs[0].to_affine(),
+        );
+
+        lhs == rhs
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::kzg::setup;
+    use rand::{rngs::SmallRng, Rng, SeedableRng};
+
+    const RNG_SEED: [u8; 32] = [69; 32];
+
+    fn test_setup<const MAX_COEFFS: usize>(rng: &mut SmallRng) -> KZGParams {
+        let s: Scalar = rng.gen::<u64>().into();
+        setup(s, MAX_COEFFS)
+    }
+
+    fn test_participants(params: &'_ KZGParams) -> (KZGProver<'_>, KZGVerifier<'_>) {
+        let prover = KZGProver::new(params);
+        let verifier = KZGVerifier::new(params);
+
+        (prover, verifier)
+    }
+
+    // never returns zero polynomial
+    fn random_polynomial(rng: &mut SmallRng, min_coeffs: usize, max_coeffs: usize) -> Polynomial {
+        let num_coeffs = rng.gen_range(min_coeffs..max_coeffs);
+        let mut coeffs = vec![Scalar::zero(); max_coeffs];
+
+        for coeff in coeffs.iter_mut().take(num_coeffs) {
+            *coeff = rng.gen::<u64>().into();
+        }
+
+        let mut poly = Polynomial::new_from_coeffs(coeffs, num_coeffs - 1);
+        poly.shrink_degree();
+        poly
+    }
+
+    fn assert_verify_poly(
+        verifier: &KZGVerifier,
+        commitment: &KZGCommitment,
+        polynomial: &Polynomial,
+    ) {
+        assert!(
+            verifier.verify_poly(commitment, polynomial),
+            "verify_poly failed for commitment {:#?} and polynomial {:#?}",
+            commitment,
+            polynomial
+        );
+    }
+
+    fn assert_verify_poly_fails(
+        verifier: &KZGVerifier,
+        commitment: &KZGCommitment,
+        polynomial: &Polynomial,
+    ) {
+        assert!(
+            !verifier.verify_poly(commitment, polynomial),
+            "expected verify_poly to fail for commitment {:#?} and polynomial {:#?} but it didn't",
+            commitment,
+            polynomial
+        );
+    }
+
+    fn assert_verify_eval(
+        verifier: &KZGVerifier,
+        point: (Scalar, Scalar),
+        commitment: &KZGCommitment,
+        witness: &KZGWitness,
+    ) {
+        assert!(
+            verifier.verify_eval(point, commitment, witness),
+            "verify_eval failed for point {:#?}, commitment {:#?}, and witness {:#?}",
+            point,
+            commitment,
+            witness
+        );
+    }
+
+    fn assert_verify_eval_fails(
+        verifier: &KZGVerifier,
+        point: (Scalar, Scalar),
+        commitment: &KZGCommitment,
+        witness: &KZGWitness,
+    ) {
+        assert!(!verifier.verify_eval(point, commitment, witness), "expected verify_eval to fail for for point {:#?}, commitment {:#?}, and witness {:#?}, but it didn't", point, commitment, witness);
+    }
+
+    #[test]
+    fn test_basic() {
+        let mut rng = SmallRng::from_seed(RNG_SEED);
+        let params = test_setup::<12>(&mut rng);
+
+        let (prover, verifier) = test_participants(&params);
+
+        let polynomial = random_polynomial(&mut rng, 2, 12);
+        let commitment = prover.commit(&polynomial);
+
+        assert_verify_poly(&verifier, &commitment, &polynomial);
+        assert_verify_poly_fails(&verifier, &commitment, &random_polynomial(&mut rng, 2, 12));
+    }
+
+    fn random_field_elem_neq(val: Scalar) -> Scalar {
+        let mut rng = SmallRng::from_seed(RNG_SEED);
+        let mut v: Scalar = rng.gen::<u64>().into();
+        while v == val {
+            v = rng.gen::<u64>().into();
+        }
+
+        v
+    }
+
+    #[test]
+    fn test_modify_single_coeff() {
+        let mut rng = SmallRng::from_seed(RNG_SEED);
+        let params = test_setup::<8>(&mut rng);
+
+        let (prover, verifier) = test_participants(&params);
+
+        let polynomial = random_polynomial(&mut rng, 3, 8);
+        let commitment = prover.commit(&polynomial);
+
+        let mut modified_polynomial = polynomial.clone();
+        let new_coeff = random_field_elem_neq(modified_polynomial.coeffs[2]);
+        modified_polynomial.coeffs[2] = new_coeff;
+
+        assert_verify_poly(&verifier, &commitment, &polynomial);
+        assert_verify_poly_fails(&verifier, &commitment, &modified_polynomial);
+    }
+
+    #[test]
+    fn test_eval_basic() {
+        let mut rng = SmallRng::from_seed(RNG_SEED);
+        let params = test_setup::<13>(&mut rng);
+
+        let (prover, verifier) = test_participants(&params);
+
+        let polynomial = random_polynomial(&mut rng, 5, 13);
+        let commitment = prover.commit(&polynomial);
+
+        let x: Scalar = rng.gen::<u64>().into();
+        let y = polynomial.eval(x);
+
+        let witness = prover.create_witness(&polynomial, (x, y)).unwrap();
+        assert_verify_eval(&verifier, (x, y), &commitment, &witness);
+
+        let y_prime = random_field_elem_neq(y);
+        assert_verify_eval_fails(&verifier, (x, y_prime), &commitment, &witness);
+
+        // test degree 1 edge case
+        let mut coeffs = vec![Scalar::zero(); 13];
+        coeffs[0] = 3.into();
+        coeffs[1] = 1.into();
+        let polynomial = Polynomial::new(coeffs);
+
+        let commitment = prover.commit(&polynomial);
+        let witness = prover
+            .create_witness(&polynomial, (1.into(), 4.into()))
+            .unwrap();
+        assert_verify_eval(&verifier, (1.into(), 4.into()), &commitment, &witness);
+        assert_verify_eval_fails(&verifier, (1.into(), 5.into()), &commitment, &witness);
+    }
 }
