+++ node/src/lib.rs
@@ -1,1 +1,3 @@
+#![deny(warnings, unsafe_code, unused)]
+
 pub mod kzg;
